+++ src/mcp_evals/models/__init__.py
@@ -1,3 +1,11 @@
+from mcp_evals.models.server import ServerConfig
 from mcp_evals.models.spec import EvalCase, EvalSuite, ToolCatalog, ToolDefinition
 
-__all__ = ["EvalCase", "EvalSuite", "ToolCatalog", "ToolDefinition"]
+__all__ = [
+    "EvalCase",
+    "EvalSuite",
+    "ServerConfig",
+    "ToolCatalog",
+    "ToolDefinition",
+]
+